created basic working video mode, still need opx integration
