+++ qualang_tools/plot/video_mode/video_mode.py
@@ -1,0 +1,316 @@
+import dash
+from dash import dcc, html
+from dash.dependencies import Input, Output
+import logging
+from .plotly_tools import xarray_to_plotly
+
+
+__all__ = ["VideoMode"]
+
+
+# Configure logging
+logging.basicConfig(level=logging.DEBUG)
+
+
+class VideoMode:
+    def __init__(self, data_acquirer, integration_time=10e-6):
+        self.data_generator = data_acquirer
+        self.integration_time = integration_time
+        self.paused = False
+        self.iteration = 0
+
+        self.app = dash.Dash(__name__, title="Video Mode")
+        self.create_layout()
+
+    def _create_axis_layout(self, axis: str):
+        xy = axis.lower()
+        XY = axis.upper()
+        return html.Div(
+            [
+                html.Label(XY, style={"text-align": "left"}),
+                html.Div(  # span
+                    [
+                        html.Label(
+                            "Span:",
+                            style={
+                                "text-align": "right",
+                                "white-space": "nowrap",
+                                "margin-left": "15px",
+                                "margin-right": "5px",
+                            },
+                        ),
+                        dcc.Input(
+                            id=f"{xy}-span",
+                            type="number",
+                            value=getattr(self.data_generator, f"{xy}_span"),
+                            min=0.01,
+                            max=getattr(self.data_generator, f"{xy}_span") * 2,
+                            debounce=True,
+                            style={
+                                "width": "55px",
+                                "text-align": "right",
+                            },
+                        ),
+                        html.Label(
+                            "V",
+                            style={
+                                "text-align": "left",
+                                "white-space": "nowrap",
+                                "margin-left": "3px",
+                            },
+                        ),
+                    ],
+                    style={"display": "flex", "margin-bottom": "10px"},
+                ),
+                html.Div(  # Points
+                    [
+                        html.Label(
+                            "Points:",
+                            style={
+                                "text-align": "right",
+                                "white-space": "nowrap",
+                                "margin-left": "15px",
+                                "margin-right": "5px",
+                            },
+                        ),
+                        dcc.Input(
+                            id=f"{xy}-points",
+                            type="number",
+                            value=getattr(self.data_generator, f"{xy}_points"),
+                            min=1,
+                            max=501,
+                            step=1,
+                            debounce=True,
+                            style={
+                                "width": "40px",
+                                "text-align": "right",
+                            },
+                        ),
+                    ],
+                    style={"display": "flex", "margin-bottom": "10px"},
+                ),
+                html.Div(  # Offset
+                    [
+                        html.Label(
+                            "Offset:",
+                            style={
+                                "text-align": "right",
+                                "white-space": "nowrap",
+                                "margin-left": "15px",
+                                "margin-right": "5px",
+                            },
+                        ),
+                        dcc.Input(
+                            id=f"{xy}-offset",
+                            type="number",
+                            value=getattr(self.data_generator, f"{xy}_offset_parameter").get(),
+                            debounce=True,
+                            style={
+                                "width": "55px",
+                                "text-align": "right",
+                            },
+                        ),
+                        html.Label(
+                            "V",
+                            style={
+                                "text-align": "left",
+                                "white-space": "nowrap",
+                                "margin-left": "3px",
+                            },
+                        ),
+                    ],
+                    style={"display": "flex", "margin-bottom": "10px"},
+                ),
+            ],
+            style={"display": "flex", "flex-direction": "row", "flex-wrap": "wrap"},
+        )
+
+    def create_layout(self):
+        self.fig = xarray_to_plotly(self.data_generator.xarr)
+
+        # Modify the layout with CSS to left-align and adjust input size
+        self.app.layout = html.Div(
+            [
+                html.Div(  # Settings
+                    [
+                        html.Header(
+                            "Video mode", style={"font-size": 32, "font-weight": "bold", "margin-bottom": "15px"}
+                        ),
+                        html.Div(  # Pause + iteration
+                            [
+                                html.Button(
+                                    "Pause",
+                                    id="pause-button",
+                                    n_clicks=0,
+                                    style={"width": "20%", "min-width": "65px"},
+                                ),
+                                html.Div(
+                                    id="iteration-output",
+                                    children="Iteration: 0",
+                                    style={"margin-left": "15px"},
+                                ),
+                            ],
+                            style={
+                                "display": "flex",
+                                "flex-direction": "row",
+                                "align-items": "center",
+                                "margin-bottom": "20px",
+                            },
+                        ),
+                        html.Div(  # Integration + Averages
+                            [
+                                html.Div(  # Integration
+                                    [
+                                        html.Label(
+                                            "Integration Time:",
+                                            style={
+                                                "text-align": "right",
+                                                "white-space": "nowrap",
+                                                "margin-right": "5px",
+                                            },
+                                        ),
+                                        dcc.Input(
+                                            id="integration-time",
+                                            type="number",
+                                            value=self.integration_time * 1e6,
+                                            min=1,
+                                            debounce=True,
+                                            style={"width": "40px", "text-align": "right"},
+                                        ),  # Integration time in microseconds
+                                        html.Label(
+                                            "µs",
+                                            style={
+                                                "text-align": "left",
+                                                "white-space": "nowrap",
+                                                "margin-left": "3px",
+                                            },
+                                        ),
+                                    ],
+                                    style={"display": "flex", "margin-bottom": "10px"},
+                                ),
+                                html.Div(
+                                    [
+                                        html.Label(
+                                            "Averages:",
+                                            style={
+                                                "text-align": "left",
+                                                "white-space": "nowrap",
+                                                "margin-left": "15px",
+                                                "margin-right": "5px",
+                                            },
+                                        ),
+                                        dcc.Input(
+                                            id="num-averages",
+                                            type="number",
+                                            value=self.integration_time * 1e6,
+                                            min=1,
+                                            step=1,
+                                            debounce=True,
+                                            style={"width": "40px"},
+                                        ),  # Integration time in microseconds
+                                    ],
+                                    style={"display": "flex", "margin-bottom": "10px"},
+                                ),
+                            ],
+                            style={
+                                "display": "flex",
+                                "flex-direction": "row",
+                                "flex-wrap": "wrap",
+                            },
+                        ),
+                        self._create_axis_layout("x"),
+                        self._create_axis_layout("y"),
+                    ],
+                    style={"width": "40%", "margin": "auto"},
+                ),
+                dcc.Graph(
+                    id="live-heatmap", figure=self.fig, style={"width": "55%", "height": "100%", "min-width": "500px"}
+                ),
+                dcc.Interval(id="interval-component", interval=self.update_interval, n_intervals=0),
+            ],
+            style={"display": "flex", "flex-direction": "row", "height": "100%", "flex-wrap": "wrap"},
+        )
+        logging.debug("Dash layout created")
+        self.add_callbacks()
+
+    def clear_data(self):
+        """Clears data history and resets averages."""
+        self.data_generator.data_history.clear()
+        logging.debug("Cleared all averages and data history.")
+
+    @property
+    def update_interval(self):
+        return self.integration_time * self.data_generator.x_points * self.data_generator.y_points * 1000
+
+    def add_callbacks(self):
+        @self.app.callback(
+            Output("pause-button", "children"),
+            [Input("pause-button", "n_clicks")],
+        )
+        def toggle_pause(n_clicks):
+            self.paused = not self.paused
+            logging.debug(f"Paused: {self.paused}")
+            return "Resume" if self.paused else "Pause"
+
+        @self.app.callback(
+            [
+                Output("live-heatmap", "figure"),
+                Output("iteration-output", "children"),
+                Output("interval-component", "interval"),
+            ],
+            [
+                Input("interval-component", "n_intervals"),
+                Input("integration-time", "value"),
+                Input("num-averages", "value"),
+                Input("x-span", "value"),
+                Input("y-span", "value"),
+                Input("x-points", "value"),
+                Input("y-points", "value"),
+                Input("x-offset", "value"),
+                Input("y-offset", "value"),
+            ],
+        )
+        def update_heatmap(
+            n_intervals, integration_time, num_averages, x_span, y_span, x_points, y_points, x_offset, y_offset
+        ):
+            if self.paused:
+                logging.debug(f"Updates paused at iteration {self.iteration}")
+                return self.fig, f"Iteration: {self.iteration}", self.update_interval
+
+            attrs = [
+                {"obj": self, "attr": "integration_time", "new": integration_time / 1e6},
+                {"obj": self.data_generator, "attr": "num_averages", "new": num_averages},
+                {"obj": self.data_generator, "attr": "x_span", "new": x_span},
+                {"obj": self.data_generator, "attr": "y_span", "new": y_span},
+                {"obj": self.data_generator, "attr": "x_points", "new": x_points},
+                {"obj": self.data_generator, "attr": "y_points", "new": y_points},
+                {"obj": self.data_generator.x_offset_parameter, "attr": "latest_value", "new": x_offset},
+                {"obj": self.data_generator.y_offset_parameter, "attr": "latest_value", "new": y_offset},
+            ]
+
+            logging.debug(f"***Updating heatmap at iteration {n_intervals}")
+            for attr in attrs:
+                attr["old"] = getattr(attr["obj"], attr["attr"])
+
+                if attr["old"] == attr["new"]:
+                    continue
+
+                logging.debug(f"Updating {attr['attr']} from {attr['old']} to {attr['new']}")
+
+                self.clear_data()
+
+                if attr["attr"] in ["x_offset", "y_offset"]:
+                    attr["obj"].set(attr["new"])
+                else:
+                    setattr(attr["obj"], attr["attr"], attr["new"])
+
+            updated_xarr = self.data_generator.update_data()
+            self.fig = xarray_to_plotly(updated_xarr)
+
+            # Increment iteration counter and update frontend
+            self.iteration += 1
+            return self.fig, f"Iteration: {self.iteration}", self.update_interval
+
+    def run(self):
+        logging.debug("Starting Dash server")
+        self.app.run_server(debug=True)
